#186: Move assertThat from impl classes to static util AssertionsForStatistics.
