+++ src/test/java/com/commercetools/sync/commons/asserts/statistics/ProductSyncStatisticsAssert.java
@@ -1,25 +1,13 @@
-package com.commercetools.sync.commons.asserts.syncStatistic;
+package com.commercetools.sync.commons.asserts.statistics;
 
 import com.commercetools.sync.products.helpers.ProductSyncStatistics;
 
-import javax.annotation.Nonnull;
 import javax.annotation.Nullable;
 
-public class ProductSyncStatisticsAssert
+public final class ProductSyncStatisticsAssert
     extends AbstractSyncStatisticsAssert<ProductSyncStatisticsAssert, ProductSyncStatistics> {
 
-    private ProductSyncStatisticsAssert(@Nullable final ProductSyncStatistics actual) {
+    ProductSyncStatisticsAssert(@Nullable final ProductSyncStatistics actual) {
         super(actual, ProductSyncStatisticsAssert.class);
-    }
-
-    /**
-     * Create assertion for {@link ProductSyncStatistics}.
-     *
-     * @param statistics the actual value.
-     * @return the created assertion object.
-     */
-    @Nonnull
-    public static ProductSyncStatisticsAssert assertThat(@Nullable final ProductSyncStatistics statistics) {
-        return new ProductSyncStatisticsAssert(statistics);
     }
 }
